udev-extraconf: Allow bypass of systemd detection
The udev-extraconf mount.sh script checks to see if /sbin/init is linked
to systemd to to determine if systemd-mount should be used. The default
behavior of the overlayfs-etc class replaces init with a setup script,
breaking this detection.

This change adds a USE_SYSTEMD_MOUNT variable that will force use of
systemd-mount when set to "1" and mount if "0". If empty, the mount
script will use the existing detection mechanism at runtime.